Extracted virtualenv deployment to second script
The virtual environment should be deployed and activated in the current shell,
otherwise the user needs to re-activate the environment once the script
finishes.  Therefore, we provide a second script, deploy_virtualenv.sh,
which is to be sourced and not run in a subshell.

Once in the vrtual environment, the other components should be deployed in a
subshell, so that errors can be handled gracefully.  Therefore, the deploy.sh
script should be run directly and not sourced.